Add footer to other pages besides the landing page
