+++ civictechprojects/templates/new_index.html
@@ -64,7 +64,7 @@
         })
         : {};
 
-      window.DLAB_FOOTER_LINKS = '{{footerLinks}}'
+      window.DLAB_FOOTER_LINKS = '{{footerLinks}}';
 
       window.DLAB_MESSAGES = [];
       {% for message in messages %}
